2110415-soft-def-sys/activity10-edge-computing/sds-edge-computing/Activity2/server/requirements.txt
ultralytics==8.3.70
torch==2.5.0
torchvision==0.20
flask
flask_cors
boto3
python-dotenv
